deleted esbuild, commented on my scripts and moved scripts from functions into functions/payment and functions/menu and so on. updated paths in scripts and serverless.yml
