+++ packages/baan-thai-backend/functions/menu/deleteMenuItem.js
