+++ packages/baan-thai-backend/functions/order/PostCreateOrder/postCreateOrder.mjs
@@ -2,9 +2,9 @@ import middy from '@middy/core'
 import httpJsonBodyParser from '@middy/http-json-body-parser'
 import { sendResponse } from '../../../responses/response.mjs'
 import { errorHandler } from '../../../middlewares/errorHandler.mjs'
-import { validateOrder } from '../../../middlewares/validateOrder.mjs'
+import { validateOrder } from '../../../middlewares/order/validateOrder.mjs'
 import { addOrder } from '../../../services/orders.mjs'
-import { queryMenuItem } from '../../queryMenuItem.mjs'
+import { queryMenuItem } from '../../menu/queryMenuItem.mjs'
 
 export const handler = middy(async (event) => {
   const body = event.body;
