+++ packages/baan-thai-backend/serverless.yml
@@ -1,27 +1,10 @@
 service: baan-thai-backend
 
-
 useDotenv: true
 
 plugins:
-  - serverless-esbuild
   - serverless-dotenv-plugin
   - serverless-offline
-
-custom:
-  esbuild:
-    bundle: true
-    minify: false
-    sourcemap: true
-    target: node20
-    platform: node
-    resolveExtensions:
-      - .mjs
-      - .js
-      - .ts
-    external:
-      - "@aws-sdk/client-dynamodb"
-      - "@aws-sdk/lib-dynamodb"
 
 provider:
   name: aws
@@ -72,35 +55,35 @@ functions:
 
   #  -------------------MENU---------------------
   getMenu: # get all meals from the menu
-    handler: functions/getMenu.handler
+    handler: functions/menu/getMenu.handler
     events:
       - httpApi:
           path: /api/menu
           method: GET
 
   getProduct: # gets a single menu item
-    handler: functions/getProduct.handler
+    handler: functions/menu/getProduct.handler
     events:
       - httpApi:
           path: /api/menu/{prodId}
           method: GET
 
   createMenuItem: # add a new meal to menu
-    handler: functions/createMenuItem.handler
+    handler: functions/menu/createMenuItem.handler
     events:
       - httpApi:
           path: /api/menu
           method: POST
 
   updateMenuItem: # change a meal in menu
-    handler: functions/updateMenuItem.handler
+    handler: functions/menu/updateMenuItem.handler
     events:
       - httpApi:
           path: /api/menu/{prodId}
           method: PUT
 
   deleteMenuItem: # removes menu product (menu item === a meal)
-    handler: functions/deleteMenuItem.handler
+    handler: functions/menu/deleteMenuItem.handler
     events:
       - httpApi:
           path: /api/menu/{prodId}
@@ -156,6 +139,13 @@ functions:
           path: /api/orders/{orderId}
           method: PUT
 
+  deleteOrder: # deletes an order completely
+    handler: functions/order/DeleteOrder/deleteOrder.handler
+    events:
+      - httpApi:
+          path: /api/orders/{orderId}
+          method: DELETE
+
   #  -------------------NOTIFICATIONS---------------------
   getNotifications: # gets all notifications for a user
     handler: functions/notifications/GetNotifications/getNotifications.handler
@@ -194,24 +184,25 @@ functions:
   #         path: /api/orders/{orderId}/messages
   #         method: GET
 
-  # TODO: Skapa payment handlers
-  # #  -------------------PAYMENTS---------------------
-  # createPayment: # create payment for order
-  #   handler: functions/createPayment.handler
-  #   events:
-  #     - httpApi:
-  #         path: /api/orders/{orderId}/payment
-  #         method: POST
+  #  -------------------PAYMENTS---------------------
+  processPayment: # process payment for order
+    handler: functions/payment/processPayment.handler
+    events:
+      - httpApi:
+          path: /api/orders/{orderId}/payment
+          method: POST
 
-  # getPayment: # get payment status
-  #   handler: functions/getPayment.handler
+  # EJ FÄRDIG - getPayment: get payment status
+  # getPayment:
+  #   handler: functions/payment/getPayment.handler
   #   events:
   #     - httpApi:
   #         path: /api/orders/{orderId}/payment
   #         method: GET
 
-  # updatePaymentStatus: # update payment status (webhook from provider)
-  #   handler: functions/updatePaymentStatus.handler
+  # EJ FÄRDIG - updatePaymentStatus: webhook from payment provider
+  # updatePaymentStatus:
+  #   handler: functions/payment/updatePaymentStatus.handler
   #   events:
   #     - httpApi:
   #         path: /api/payments/{paymentId}/status
